requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
streamlit==1.28.0
openai==1.3.0
tavily-python==0.3.0
google-search-results==2.4.2
plotly==5.17.0
Pillow==10.0.0